requirements-COMPLETE.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
spacy==3.7.2
pytesseract==0.3.10
pdfplumber==0.10.3
python-docx==0.8.11
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
boto3==1.28.85
aiofiles==23.2.1
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
